Added ability to stop the bot and cause it to leave a channel
The bot now holds one copy of the speaker class and can use it to keep track of the current state and call the stop method to stop the speaker from downloading and playing the sound.

I moved the url from the Speaker constructor to the load method to allow the speaker class to be created before the url is known.